requirements.txt
Django==3.1.7
python-dotenv==0.15.0
psycopg2-binary==2.8.6
phonenumbers==8.12.15
oauthlib==3.1.0
djangorestframework==3.12.2
requests==2.25.1
six==1.15.0
django-mptt==0.11.0
text-unidecode==1.3
Pillow==8.0.1
django-cors-headers==3.6.0
python-telegram-bot==13.1